Fix minor issues revealed with -Weverythinig (#170)
Flags on apple clang 12.0.0:
```
-Wall -Weverything -Wno-c++98-compat -Wno-c++98-compat-pedantic
-Wno-missing-noreturn -Wno-global-constructors -Wno-padded
-Wno-documentation-unknown-command -pedantic -Wno-exit-time-destructors
-Wno-sign-conversion -Wno-weak-vtables -Wno-float-equal
-Wno-used-but-marked-unused -fdiagnostics-color=always
```

+++ src/random/cuda/RngStateStore.hh
@@ -24,7 +24,7 @@ class RngStateStore
     RngStateStore() = default;
 
     // Construct with the number of RNG states
-    explicit RngStateStore(size_type size, unsigned long host_seed = 12345u);
+    explicit RngStateStore(size_type size, unsigned int host_seed = 12345u);
 
     //! Number of states
     size_type size() const { return data_.size(); }
